+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 ## [Unreleased][]
 
+## [Release][1.7.3] [1.7.3][1.7.2-1.7.3] - 2022-03-25
+
+### Added 
+- New Keywords Get Rectangle Bounding From Element
+
 ## [Release][1.7.2] [1.7.2][1.7.1-1.7.2] - 2022-02-13
 
 ### Added